% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/averageReps.R
\name{avergeReps}
\alias{avergeReps}
\title{averageReps}
\usage{
avergeReps(geneExpr, Reps)
}
\arguments{
\item{geneExpr}{double}

\item{Reps}{int}
}
\value{
double
}
\description{
averages expr values within replicate groups
}
\examples{

geneExpr <- c(1, 5, 3, 4, 5, 8)
reps <-  c(2, 3, 1)
avergeReps(geneExpr = geneExpr, Reps = reps)
}
