% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeLaplacianEmbedding.R
\name{computeLaplacianEmbedding}
\alias{computeLaplacianEmbedding}
\title{computeLaplacianEmbedding}
\usage{
computeLaplacianEmbedding(takens)
}
\arguments{
\item{takens}{double}
}
\value{
double
}
\description{
Compute the Laplacian Embedding about the merged Takens embedding
}
