% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computePersistence2.R
\name{computePersistence2.0}
\alias{computePersistence2.0}
\title{computePersistence2.0}
\usage{
computePersistence2.0(data, minLag = 1, maxLag = 6, cores = 1)
}
\arguments{
\item{data}{int}

\item{minLag}{int}

\item{maxLag}{int}

\item{cores}{int}
}
\value{
int
}
\description{
Returns the Lag with the Top Persistence Score
}
