% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detrend.R
\name{detrend}
\alias{detrend}
\title{Linearly Detrends a \code{data.frame} of Gene Expression Time-Series}
\usage{
detrend(data)
}
\arguments{
\item{data}{a \code{data.frame} of \code{numeric} gene expression over time (row = genes \emph{x} col = ZT times).}
}
\value{
a detrended \code{data.frame} of \code{numeric} gene expression over time (row = genes \emph{x} col = ZT times).
}
\description{
Detrends a \code{data.frame} of gene expression over time (row = genes \emph{x} col = ZT times)
by fiting a linear model to each gene and removing the linear trend.
}
\seealso{
\code{\link{periodFinder}}
}
