% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFFT.R
\name{getFFT}
\alias{getFFT}
\title{Computes the Fast Fourier Transform of a Time-Series}
\usage{
getFFT(y, sampFreq)
}
\arguments{
\item{y}{a \code{vector} of numeric time-series expression values.}

\item{sampFreq}{a \code{vector} of numeric frequencies.}
}
\value{
a \code{data.frame} harmonic fits by frequency.
}
\description{
Takes a time-series and sample frequency and computes the fast Fourier transform of the time-series.
}
\seealso{
\code{\link[stats]{fft}}
}
