% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPersistence.R
\name{getPersistence}
\alias{getPersistence}
\title{getPersistence}
\usage{
getPersistence(TimeSeries, lag, laplacian = T)
}
\arguments{
\item{TimeSeries}{double}

\item{lag}{int}

\item{laplacian}{boolean}
}
\value{

}
\description{
Returns the Max persistence score, returns 0 if no persistence score exists
}
