% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTimeSpan.R
\name{getTimeSpan}
\alias{getTimeSpan}
\title{Get Differene Between Start Time and End Time}
\usage{
getTimeSpan(start, end)
}
\arguments{
\item{start}{double}

\item{end}{double}
}
\value{
double
}
\description{
Converts System Time Differences to
hh:mm:ss format
}
\examples{
getTimeSpan(Sys.time(),Sys.time()+600)
}
