% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputeMissingData.R
\name{imputeMissingData}
\alias{imputeMissingData}
\title{Imputates Missing Time-Points in Data}
\usage{
imputeMissingData(data)
}
\arguments{
\item{data}{a \code{data.frame} of \code{numeric} gene expression over time (row = genes \emph{x} col = ZT times) with missing values.}
}
\value{
a imputed \code{data.frame} of \code{numeric} gene expression over time (row = genes \emph{x} col = ZT times).
}
\description{
Imputes \code{numeric} values for time-points with an \code{NA} by computing the linear path between missing points
}
\examples{
a <- c(10, 12, 14, NA, 18)
b <- c(1, 2, NA, NA, 5)
data <- t(data.frame(a, b))
imputeMissingData(data)
}
\seealso{
\code{\link[imputeTS]{na_interpolation}} for imputation procedure.
}
