% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixLaplacian.R
\name{matrixLaplacian}
\alias{matrixLaplacian}
\title{Computes Laplacian Eigenmaping}
\usage{
matrixLaplacian(A)
}
\arguments{
\item{A}{a distance \code{matrix} of the n-dimensional Takens' embedding.}
}
\value{
a \code{list} of the laplacian matrix and eigenvectors.
}
\description{
Computes the Laplacian matrix and eigenvectors of the n-D Takens' embedding.
}
\seealso{
\code{\link{computeLaplacianEmbedding}}
}
