% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meanCenter.R
\name{meanCenter}
\alias{meanCenter}
\title{Mean Centers a \code{data.frame}}
\usage{
meanCenter(df)
}
\arguments{
\item{df}{a \code{data.frame} of \code{numerics}.}
}
\value{
a mean center \code{data.frame} of \code{numerics} by row.
}
\description{
Mean centers a \code{data.frame} by row.
}
