% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/movingAverageDF.R
\name{movingAverageDF}
\alias{movingAverageDF}
\title{movingAverageDF}
\usage{
movingAverageDF(data)
}
\arguments{
\item{data}{double DataFrame /Rows = Genes, Column = Sample ZT/}
}
\value{
double DataFrame /Rows = Genes, Column = Sample ZT/
}
\description{
Computes the moving average across TimeSeries in a data frame
}
