% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nullResampling.R
\name{nullResampling}
\alias{nullResampling}
\title{Generates a \code{data.frame} of Resampled Time-Series for Computing the Null Distribution}
\usage{
nullResampling(data, numExperiments = 10000)
}
\arguments{
\item{data}{a \code{data.frame} of \code{numeric} gene expression over time (row = genes \emph{x} col = ZT times).}

\item{numExperiments}{a \code{numeric} specifying the number of resampling to use in the null-distribution. Default is \code{10000}.}
}
\value{
a \code{data.frame} of \code{numeric} resampled time-series (row = numExperiments \emph{x} col = ZT times).
}
\description{
Generates a \code{data.frame} null resampling time-series as defined by the difference between consecutive points across all time-series.
For additional information about the null distribution used by TimeCycle, see TimeCycle's vignette:
\code{vignette("TimeCycle")}.
}
\seealso{
\code{\link{resampleTimeSeries}}
}
