% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{preprocess}
\alias{preprocess}
\title{Preprocess}
\usage{
preprocess(data)
}
\arguments{
\item{data}{double dataframe /Rows = Genes, Column = Sample ZT/}
}
\value{
double preprocessed dataframe /Rows = Genes, Column = Sample ZT/
}
\description{
Preprocesses the data by smoothing with an sgolay filter, mean centering, and scaling
}
