% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_acf.R
\name{preprocess_acf}
\alias{preprocess_acf}
\title{Computes Time-Series Autocovariance Function}
\usage{
preprocess_acf(data, period = 24)
}
\arguments{
\item{data}{a \code{data.frame} of \code{numeric} gene expression over time (row = genes \emph{x} col = ZT times).}

\item{period}{a \code{numeric} specifying the period of interest in hours for rhythm detection. Default is \code{24}.}
}
\value{
a smoothed \code{data.frame} of \code{numeric} gene expression covariance over time (row = genes \emph{x} col = ZT times).
}
\description{
Computing the autocovariance function from a \code{data.frame} of time-series. The resulting autocovariance
function is smoothed using a moving average defined by the period and sampling scheme.
Returns the smoothed time-series as a \code{data.frame}.
}
\seealso{
\code{\link{meanCenter}}, \code{\link{scaleTimeSeries}}
}
