% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoothDataSgolay.R
\name{preprocess_sgolay}
\alias{preprocess_sgolay}
\title{preprocess_sgolay}
\usage{
preprocess_sgolay(data)
}
\arguments{
\item{data}{double DataFrame /Rows = Genes, Column = Sample ZT/}
}
\value{
double Savitzky-Golay Filtered by Row
}
\description{
Smooth Data Using Savitzky–Golay filter
}
