% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resampleTimeSeries.R
\name{resampleTimeSeries}
\alias{resampleTimeSeries}
\title{Generates a \code{data.frame} of Resampled Time-Series}
\usage{
resampleTimeSeries(distTP, numTP, numRsmps)
}
\arguments{
\item{distTP}{a \code{data.frame} of \code{numeric} values.}

\item{numTP}{a \code{numeric} specifying the number of columns in the outputted \code{data.frame}.}

\item{numRsmps}{a \code{numeric} specifying the number of rows in the outputted \code{data.frame}.}
}
\value{
a \code{data.frame} of randomly sampled \code{numerics} time-series (row = \code{numRsamps} \emph{x} col = \code{numTP}).
}
\description{
Converts a \code{data.frame} of \code{numeric} values into a single vector and generates a random resampling with dimension
\code{numRsmps} by \code{numTP}.
}
\seealso{
\code{\link{nullResampling}}
}
