% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaleTimeSeries.R
\name{scaleTimeSeries}
\alias{scaleTimeSeries}
\title{scale Time Series}
\usage{
scaleTimeSeries(TimeSeries)
}
\arguments{
\item{TimeSeries}{double}
}
\value{
double
}
\description{
scales Time Series values to be between 0 and 1
}
