% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoothDataSgolay.R
\name{smoothData.sgolay}
\alias{smoothData.sgolay}
\title{smoothData.sgolay}
\usage{
smoothData.sgolay(data)
}
\arguments{
\item{data}{double DataFrame /Rows = Genes, Column = Sample ZT/}
}
\value{
double Savitzky-Golay Filtered by Row
}
\description{
Smooth Data Using Savitzky–Golay filter
}
