% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TimeCycle.R
\name{TimeCycle}
\alias{TimeCycle}
\title{TimeCycle}
\usage{
TimeCycle(
  data,
  repLabel,
  experiments = 10000,
  minLag = 1,
  maxLag = 6,
  cores = detectCores() - 2
)
}
\arguments{
\item{data}{double}

\item{repLabel}{int}

\item{experiments}{int}

\item{minLag}{int}

\item{maxLag}{int}

\item{cores}{int}
}
\value{

}
\description{
Main Function
}
