% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/averageReps.R
\name{averageReps}
\alias{averageReps}
\title{averageReps}
\usage{
averageReps(geneExpr, Reps)
}
\arguments{
\item{geneExpr}{double}

\item{Reps}{int}
}
\value{
double
}
\description{
averages expr values within replicate groups
}
\examples{

geneExpr <- c(1, 5, 3, 4, 5, 8)
reps <- c(2, 3, 1)
averageReps(geneExpr = geneExpr, Reps = reps)
}
