% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildTakens_ndim.R
\name{buildTakens_ndim}
\alias{buildTakens_ndim}
\title{Generates Taken Embedding}
\usage{
buildTakens_ndim(x, dim, delay = 1, indices = FALSE, as.embed = FALSE)
}
\arguments{
\item{x}{int}

\item{dim}{int}

\item{delay}{int}

\item{indices}{boolean}

\item{as.embed}{boolean}
}
\value{

}
\description{
Generates the Takens Embedding given a specified delay and dimension
}
