% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computePersistence.R
\name{computePersistence}
\alias{computePersistence}
\title{computePersistence2.0}
\usage{
computePersistence(data, minLag = 2, maxLag = 5, cores = 1, laplacian = T)
}
\arguments{
\item{data}{int}

\item{minLag}{int}

\item{maxLag}{int}

\item{cores}{int}

\item{laplacian}{boolean}
}
\value{
int
}
\description{
Returns the Lag with the Top Persistence Score
}
