% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detrend.R
\name{detrend}
\alias{detrend}
\title{Detrend}
\usage{
detrend(data)
}
\arguments{
\item{data}{double dataframe /Rows = Genes, Column = Sample ZT/}
}
\value{
double detrended dataframe /Rows = Genes, Column = Sample ZT/
}
\description{
Detrend Data by fiting a linear model and removing the linear trend
}
