% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRepAvgedDataFrame.R
\name{getRepAvgedDataFrame}
\alias{getRepAvgedDataFrame}
\title{getRepAvgedDataFrame}
\usage{
getRepAvgedDataFrame(data, repLabel)
}
\arguments{
\item{data}{double dataframe /Rows = Genes, Column = Sample ZT/}

\item{repLabel}{int}
}
\value{

}
\description{
averages expr values within replicate groups by row in a dataframe
}
