% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTimeSpan.R
\name{getTimeSpan}
\alias{getTimeSpan}
\title{Computes Algorithm Run Time}
\usage{
getTimeSpan(start, end)
}
\arguments{
\item{start}{a \code{numeric} defining the algorithm start time.}

\item{end}{a \code{numeric} defining the algorithm end time.}
}
\value{
a \code{numeric} defining the algorithm run time.
}
\description{
Converts system time difference between TimeCycle's start and end time to
hh:mm:ss format.
}
\examples{
getTimeSpan(Sys.time(), Sys.time() + 600)
}
