% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputeMissingData.R
\name{imputeMissingData}
\alias{imputeMissingData}
\title{ImputeMissingData}
\usage{
imputeMissingData(data)
}
\arguments{
\item{data}{df (Col = Genes, Row = TimePoints)}
}
\value{

}
\description{
Imputes NA in the data frame by computing the linear path between missing points
}
\examples{
a <- c(10, 12, 14, NA, 18)
b <- c(1, 2, NA, NA, 5)
data <- t(data.frame(a, b))
imputeMissingData(data)
}
