% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixLaplacian.R
\name{matrixLaplacian}
\alias{matrixLaplacian}
\title{Compute Matrix Laplacian About Takens Embedding}
\usage{
matrixLaplacian(A, plot2D = FALSE, plot3D = FALSE)
}
\arguments{
\item{A}{int}

\item{plot2D}{boolean}

\item{plot3D}{boolean}
}
\value{

}
\description{
Compute Matrix Laplacian About Takens Embedding
}
