% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/movingAverage.R
\name{movingAverage}
\alias{movingAverage}
\title{movingAverage}
\usage{
movingAverage(x, n = 1, centered = FALSE)
}
\arguments{
\item{x}{double the time series}

\item{n}{int the number of points to use in the moving average}

\item{centered}{boolean}
}
\value{
double
}
\description{
movingAverage
}
