% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/movingAverageDF.R
\name{movingAverageDF}
\alias{movingAverageDF}
\title{Computes the Moving Average By Row in a \code{data.frame} of Time-Series}
\usage{
movingAverageDF(data)
}
\arguments{
\item{data}{a \code{data.frame} of \code{numeric} time-series expression values.}
}
\value{
a \code{data.frame} containing the smoothed \code{numeric} moving average time-series expression values by row.
}
\description{
Computes the moving average about each time-series in a \code{data.frame}.
}
\seealso{
\code{\link{movingAverage}} for parameter definitions
}
