% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nullResampling.R
\name{nullResampling}
\alias{nullResampling}
\title{Null Resampling}
\usage{
nullResampling(data, numExperiments)
}
\arguments{
\item{data}{double dataframe /Rows = Genes, Column = Sample ZT/}

\item{numExperiments}{int number of Resamplings to perform}
}
\value{
double dataframe /Rows = Genes, Column = Sample ZT/
}
\description{
Generates a null resampling as defined by rhe difference between consecutive points across all time series
}
