% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/periodFinder.R
\name{periodFinder}
\alias{periodFinder}
\title{periodFinder}
\usage{
periodFinder(data, plotFit = F)
}
\arguments{
\item{data}{double dataframe (row = genes, col = timepoints)}

\item{plotFit}{boolean}
}
\value{

}
\description{
Computes the period, amp, and phase of a dataframe of time series
}
