% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_acf.R
\name{preprocess_acf}
\alias{preprocess_acf}
\title{Preprocessing AutoCorrelation Function}
\usage{
preprocess_acf(data, period = 24)
}
\arguments{
\item{data}{dataframe of time series}

\item{period}{the period of interest in your data (i.e. default = 24)}
}
\value{

}
\description{
Take a dataframe of time-series as input and smooth oscillations by computing the autocorrelation function
and taking the moving average across time points according the defined period and sampling scheme.
Returns the smoothed time-series as a dataframe.
}
