% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resampleTimeSeries.R
\name{resampleTimeSeries}
\alias{resampleTimeSeries}
\title{resampleTimeSeries}
\usage{
resampleTimeSeries(distTP, numTP, numRsmps)
}
\arguments{
\item{distTP}{double}

\item{numTP}{int}

\item{numRsmps}{int}
}
\value{
int dataframe
}
\description{
Takes a dataframe of expr values and generates a resampling of the data used for parametric testing
}
