% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaleTimeSeries.R
\name{scaleTimeSeries}
\alias{scaleTimeSeries}
\title{Scales a Time-Series}
\usage{
scaleTimeSeries(TimeSeries)
}
\arguments{
\item{TimeSeries}{a \code{vector} of \code{numeric} time-series expression values.}
}
\value{
a scaled \code{vector} of \code{numeric} time-series expression values between 0  and 1.
}
\description{
Scales a time-series between 0 and 1.
}
\seealso{
\code{\link{preprocess_acf}}, \code{\link{meanCenter}}
}
