% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TimeCycle.R
\name{TimeCycle}
\alias{TimeCycle}
\title{TimeCycle}
\usage{
TimeCycle(
  data,
  repLabel,
  experiments = 10000,
  minLag = 2,
  maxLag = 5,
  period = 24,
  cores = parallel::detectCores() - 2,
  laplacian = T
)
}
\arguments{
\item{data}{double}

\item{repLabel}{int}

\item{experiments}{int}

\item{minLag}{int}

\item{maxLag}{int}

\item{period}{int}

\item{cores}{int}

\item{laplacian}{boolean}
}
\value{

}
\description{
Main Function
}
