% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TimeCycle.R
\name{TimeCycle}
\alias{TimeCycle}
\title{Main TimeCycle Function}
\usage{
TimeCycle(
  data,
  repLabel,
  resamplings = 10000,
  minLag = 2,
  maxLag = 5,
  period = 24,
  cores = parallel::detectCores() - 2,
  laplacian = T,
  linearTrend = F
)
}
\arguments{
\item{data}{a \code{data.frame} of numeric gene expression over time (row = genes \emph{x} col = ZT times).}

\item{repLabel}{a \code{vector} defining the number of replicates at each time points.}

\item{resamplings}{a \code{numeric} specifying the number of resamplings to use in the null-distribution. Default is \code{10000}.}

\item{minLag}{a \code{numeric} specifying the min lag to check in the 3-D embedding. Default is \code{2}.}

\item{maxLag}{a \code{numeric} specifying the max lag to check in the 3-D embedding. Default is  \code{5}.}

\item{period}{a \code{numeric} specifying the period of interest in hours for rhythm detection. Default is \code{24}.}

\item{cores}{a \code{numeric} specifying the number of parallel cores to use. Default number of cores is \code{parallel::detectedCores() - 2}.}

\item{laplacian}{a \code{logical} scalar. Should the Laplacian Eigenmaps be used for dimensionality reduction? Default \code{TRUE}}

\item{linearTrend}{a \code{logical} scalar. Should TimeCycle Prioritize detecting linear trending signals? Default \code{FALSE}. Not recommended to change from default \code{FALSE} - will increases false positives rate. See vignette("TimeCycle") for more details.}
}
\value{
a tidy \code{data.frame} of processed results for each gene:
\tabular{lcccccc}{
 \strong{sampleName} \tab  \strong{perScore} \tab \strong{pVals} \tab \strong{pVals.adj} \tab \strong{Period.in.Hours} \tab \strong{Amp} \tab \strong{Phase.In.Hours} \cr
 the gene name \tab  the median persistence score across all lags (min to max)  \tab raw empirical \emph{p}-value \tab FDR adjusted \emph{p}-value \tab period (h) \tab amplitude \tab phase (h) \cr
}
}
\description{
Main function of the \pkg{TimeCycle} package used for detecting rhythmic signals in time-series gene expression sets.
For additional help with parameter selection, see TimeCycle's vignette:
\code{vignette("TimeCycle")}.
}
\examples{
# use built in zhang2014 data set sampled every
# 2 hours for 48 hours (i.e. 24 time points with 1 replicate each).
# Search for genes with period of 24 hours.

#set seed for reproducibility with random variables in example usage
#> set.seed(1234)

#> TimeCycleResults <- TimeCycle(data = zhang2014,
#>                               repLabel = rep(1,24),
#>                               period = 24)

# Check number of genes with FDR < 0.05 and period between 22 to 26 hours.
#> library(tidyverse)

#> TimeCycleResults \%>\%
#>    filter(22 < Period.in.Hours & Period.in.Hours < 26) \%>\%
#>    filter(pVals.adj < 0.05) \%>\%
#>    glimpse()

}
\references{
{
\itemize{
   \item{A pre-print version of TimeCycle is available on BioRxiv at \url{https://doi.org/10.1101/2020.11.19.389981}}
   }
   \subsection{TDA Package References}{
   \itemize{
     \item Wadhwa RR, Williamson DFK, Dhawan A, Scott JG. (2018). "TDAstats: R pipeline for computing persistent homology in topological data analysis." \emph{Journal of Open Source Software}. 2018; 3(28): 860. doi:\href{https://doi.org/10.21105/joss.00860}{[10.21105/joss.00860]}
     \item Bauer U. (2019). "Ripser: Efficient computation of Vietoris-Rips persistence barcodes." \emph{arXiv}: 1908.02518.
   }


   }
}
}
